fix: extract NavBarVariant to separate file to prevent circular imports
