+++ web/src/components/navbar/NavBarTypes.ts
@@ -1,0 +1,9 @@
+export enum NavBarVariant {
+  LOGO_ONLY = "LOGO_ONLY",
+  LOGO_WITH_TEXT = "LOGO_WITH_TEXT",
+  MINIMAL_WITH_LOGIN = "MINIMAL_WITH_LOGIN",
+  FULL_LANDING = "FULL_LANDING",
+  MINIMAL_WITH_DISABLED_DROPDOWN = "MINIMAL_WITH_DISABLED_DROPDOWN",
+  FULL_AUTHENTICATED = "FULL_AUTHENTICATED",
+  FULL_GUEST = "FULL_GUEST",
+}
