+++ web/src/components/njwds-layout/PageSkeleton.tsx
@@ -2,7 +2,8 @@ import { BetaBar } from "@/components/BetaBar";
 import { LegalMessage } from "@/components/LegalMessage";
 import { PageFooter } from "@/components/PageFooter";
 import { SkipToMainContent } from "@/components/SkipToMainContent";
-import { NavBar, NavBarVariant } from "@/components/navbar/NavBar";
+import { NavBar } from "@/components/navbar/NavBar";
+import { NavBarVariant } from "@/components/navbar/NavBarTypes";
 import { Banner } from "@/components/njwds/Banner";
 import { AuthContext } from "@/contexts/authContext";
 import { ROUTES } from "@/lib/domain-logic/routes";
